Add cache_ttl param to AsymmetricSignatureVerifier
This new, optional parameter allows setting the cache TTL for the
underlying `JwksFetcher`. This allows caching the JWK set for more (or
less) time than the default 600 seconds.

`AsymmetricSignatureVerifier` had to be moved below `JwksFetcher`
because it now references it, so the latter has to be defined earlier in
the file.